// migrations/snapshots/30dbf9ea60f14ba627e79519f0a0e37efa2e2eac21f2c33cdf9e0445e5d65b36/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Contact: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly phoneNumber: CodecTypes['pg/text@1']['output'];
    };
    readonly Customer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly rate: CodecTypes['pg/float8@1']['output'];
    };
    readonly Delivery: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly routeId: CodecTypes['pg/int4@1']['output'];
      readonly locationId: CodecTypes['pg/int4@1']['output'];
      readonly contactId: CodecTypes['pg/int4@1']['output'] | null;
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly notes: CodecTypes['pg/text@1']['output'];
      readonly tankDetails: CodecTypes['pg/text@1']['output'];
    };
    readonly Docket: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly volume: CodecTypes['pg/int4@1']['output'];
      readonly batchNumber: CodecTypes['pg/text@1']['output'];
      readonly repName: CodecTypes['pg/text@1']['output'];
      readonly repSignature: CodecTypes['pg/text@1']['output'];
      readonly deliveryId: CodecTypes['pg/int4@1']['output'];
    };
    readonly InvoiceEmail: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly emailAddress: CodecTypes['pg/text@1']['output'];
    };
    readonly Location: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly address: CodecTypes['pg/text@1']['output'];
    };
    readonly Route: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly assignedUserId: CodecTypes['pg/int4@1']['output'];
      readonly date: CodecTypes['pg/date-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly role: 'DRIVER' | 'MANAGER';
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Contact: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly phoneNumber: CodecTypes['pg/text@1']['input'];
    };
    readonly Customer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly rate: CodecTypes['pg/float8@1']['input'];
    };
    readonly Delivery: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly routeId: CodecTypes['pg/int4@1']['input'];
      readonly locationId: CodecTypes['pg/int4@1']['input'];
      readonly contactId: CodecTypes['pg/int4@1']['input'] | null;
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly notes: CodecTypes['pg/text@1']['input'];
      readonly tankDetails: CodecTypes['pg/text@1']['input'];
    };
    readonly Docket: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly volume: CodecTypes['pg/int4@1']['input'];
      readonly batchNumber: CodecTypes['pg/text@1']['input'];
      readonly repName: CodecTypes['pg/text@1']['input'];
      readonly repSignature: CodecTypes['pg/text@1']['input'];
      readonly deliveryId: CodecTypes['pg/int4@1']['input'];
    };
    readonly InvoiceEmail: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly emailAddress: CodecTypes['pg/text@1']['input'];
    };
    readonly Location: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly address: CodecTypes['pg/text@1']['input'];
    };
    readonly Route: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly assignedUserId: CodecTypes['pg/int4@1']['input'];
      readonly date: CodecTypes['pg/date-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly role: 'DRIVER' | 'MANAGER';
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly contact: {
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly phoneNumber: CodecTypes['pg/text@1']['output'];
    };
    readonly customer: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly rate: CodecTypes['pg/float8@1']['output'];
    };
    readonly delivery: {
      readonly contactId: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly locationId: CodecTypes['pg/int4@1']['output'];
      readonly notes: CodecTypes['pg/text@1']['output'];
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly routeId: CodecTypes['pg/int4@1']['output'];
      readonly tankDetails: CodecTypes['pg/text@1']['output'];
    };
    readonly docket: {
      readonly batchNumber: CodecTypes['pg/text@1']['output'];
      readonly deliveryId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly repName: CodecTypes['pg/text@1']['output'];
      readonly repSignature: CodecTypes['pg/text@1']['output'];
      readonly volume: CodecTypes['pg/int4@1']['output'];
    };
    readonly invoiceEmail: {
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly emailAddress: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
    };
    readonly location: {
      readonly address: CodecTypes['pg/text@1']['output'];
      readonly customerId: CodecTypes['pg/int4@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
    };
    readonly route: {
      readonly assignedUserId: CodecTypes['pg/int4@1']['output'];
      readonly date: CodecTypes['pg/date-string@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly role: 'DRIVER' | 'MANAGER';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly contact: {
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly phoneNumber: CodecTypes['pg/text@1']['input'];
    };
    readonly customer: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly rate: CodecTypes['pg/float8@1']['input'];
    };
    readonly delivery: {
      readonly contactId: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly locationId: CodecTypes['pg/int4@1']['input'];
      readonly notes: CodecTypes['pg/text@1']['input'];
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly routeId: CodecTypes['pg/int4@1']['input'];
      readonly tankDetails: CodecTypes['pg/text@1']['input'];
    };
    readonly docket: {
      readonly batchNumber: CodecTypes['pg/text@1']['input'];
      readonly deliveryId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly repName: CodecTypes['pg/text@1']['input'];
      readonly repSignature: CodecTypes['pg/text@1']['input'];
      readonly volume: CodecTypes['pg/int4@1']['input'];
    };
    readonly invoiceEmail: {
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly emailAddress: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
    };
    readonly location: {
      readonly address: CodecTypes['pg/text@1']['input'];
      readonly customerId: CodecTypes['pg/int4@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
    };
    readonly route: {
      readonly assignedUserId: CodecTypes['pg/int4@1']['input'];
      readonly date: CodecTypes['pg/date-string@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
    };
    readonly user: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly role: 'DRIVER' | 'MANAGER';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;